relocate dnsdb plugin to proper path, fix imports and cleanup unnecessary code
