--- pydat5-dev/frontend/src/components/passive/index.jsx
@@ -1,13 +1,0 @@
-import React from 'react'
-import DNSDB from './dnsdb'
-
-const PassiveHander = (props) => {
-
-    return (
-        <React.Fragment>
-            <DNSDB/>
-        </React.Fragment>
-    )
-}
-
-export default PassiveHander
