+++ pydat5-dev/frontend/src/plugins/index.jsx
@@ -2,3 +2,5 @@
 // To import a plugin, place the directory of the plugin into this directory
 // and then import the directory, e.g.,
 // import './my_plugin_directory'
+
+import './dnsdb'
